+++ metatensor-core/src/io/block.rs
@@ -167,7 +167,7 @@ pub(super) fn write_single_block<W: std::io::Write + std::io::Seek>(
     values: bool,
     block: &TensorBlock,
 ) -> Result<(), Error> {
-    let options = zip::write::SimpleFileOptions::default()
+    let options = zip::write::FileOptions::default()
         .compression_method(zip::CompressionMethod::Stored)
         .large_file(true)
         .last_modified_time(zip::DateTime::from_date_and_time(2000, 1, 1, 0, 0, 0).expect("invalid datetime"));