+++ metatensor-core/src/tensor/mod.rs
@@ -1,6 +1,8 @@
-use std::collections::HashMap;
+use std::collections::{BTreeMap, HashMap};
 use std::sync::Arc;
+use std::ffi::CString;
 
+use crate::utils::ConstCString;
 use crate::TensorBlock;
 use crate::{Labels, Error};
 use crate::get_data_origin;
@@ -24,7 +26,10 @@ mod keys_to_properties;
 pub struct TensorMap {
     keys: Arc<Labels>,
     blocks: Vec<TensorBlock>,
-    // TODO: arbitrary tensor-level metadata? e.g. using `HashMap<String, String>`
+
+    info: BTreeMap<String, ConstCString>,
+    // all the keys from `info`, as C-compatible strings
+    info_keys: Vec<ConstCString>
 }
 
 fn check_labels_names(
@@ -171,6 +176,8 @@ impl TensorMap {
         Ok(TensorMap {
             keys: keys,
             blocks,
+            info: BTreeMap::new(),
+            info_keys: Vec::new(),
         })
     }
 
@@ -184,7 +191,9 @@ impl TensorMap {
 
         return Ok(TensorMap {
             keys: Arc::clone(&self.keys),
-            blocks
+            blocks,
+            info: self.info.clone(),
+            info_keys: self.info_keys.clone()
         });
     }
 
@@ -272,6 +281,33 @@ impl TensorMap {
 
         return Ok(clone);
     }
+
+    /// Get the info map associated with this `TensorMap`.
+    pub fn info(&self) -> &BTreeMap<String, ConstCString> {
+        &self.info
+    }
+
+    /// Get the list of info keys as C-compatible strings
+    pub fn info_keys_c(&self) -> &[ConstCString] {
+        &self.info_keys
+    }
+
+    /// Get the info value associated with `key` for this `TensorMap`.
+    pub fn get_info(&self, key: &str) -> Option<&ConstCString> {
+        self.info.get(key)
+    }
+
+    /// Set the info value associated with `key` for this `TensorMap`.
+    pub fn add_info(&mut self, key: &str, value: ConstCString) {
+        if !self.info.contains_key(key) {
+            self.info_keys.push(
+                ConstCString::new(
+                    CString::new(key.to_owned()).expect("invalid C string")
+                )
+            );
+        }
+        self.info.insert(key.to_owned(), value);
+    }
 }
 
 
@@ -304,6 +340,12 @@ mod tests {
             vec![block_1, block_2],
         );
         assert!(result.is_ok());
+        // also check we have an empty info block
+        let mut result = result.unwrap();
+        assert!(result.info.is_empty());
+        // also check we can set info
+        result.info.insert("key".to_string(),
+                          ConstCString::new(CString::new("value").expect("CString::new failed")));
 
         /**********************************************************************/
         let block_1 = TensorBlock::new(