+++ metatensor-torch/CHANGELOG.md
@@ -17,10 +17,24 @@ a changelog](https://keepachangelog.com/en/1.1.0/) format. This project follows
 ### Removed
 -->
 
+## [Version 0.8.3](https://github.com/metatensor/metatensor/releases/tag/metatensor-torch-v0.8.1) - 2025-12-05
+
+### Added
+
+- It is now possible to store and retrieve global metatadata about a TensorMap,
+  in the form of string key/value pairs. You can use `TensorMap::set_info`,
+  `TensorMap::get_info` and `TensorMap::info` to manipulate this information.
+
+## [Version 0.8.2](https://github.com/metatensor/metatensor/releases/tag/metatensor-torch-v0.8.1) - 2025-11-05
+
 ### Changed
 
 - We now requires at least cmake v3.22 to compile metatensor-torch
 - We now require Python >= 3.10
+
+### Added
+
+- Added support for PyTorch v2.9
 
 ## [Version 0.8.1](https://github.com/metatensor/metatensor/releases/tag/metatensor-torch-v0.8.1) - 2025-09-24